reorganize, memoize assets proxy, reference resulting files correctly
