+++ src/core/index.ts
@@ -1,6 +1,7 @@
 // most files inside of core will not need to be exported (internal use only)
 
 export * from "./utils/hooks";
+export * from "./contexts";
 export * from "./types";
 export * from "./environments/StandardEnvironment";
 export * from "./environments/KeyframeEnvironment";
