+++ src/core/loader/loaders/EnvAsset.tsx
@@ -1,15 +1,16 @@
-import { useEffect } from "react";
+import { useContext, useEffect } from "react";
 import { RGBELoader } from "three/examples/jsm/loaders/RGBELoader";
 import { UnsignedByteType } from "three";
-import { Assets } from "../../types/loading";
+import { LoadingContext } from "../../contexts";
+import { ref } from "valtio";
 
 type Props = {
   url: string;
-  assets: Assets;
 };
 
 const EnvAsset = (props: Props) => {
-  const { url, assets } = props;
+  const { url } = props;
+  const { assets } = useContext(LoadingContext);
 
   const loader = new RGBELoader();
   loader.setDataType(UnsignedByteType);
@@ -17,7 +18,7 @@ const EnvAsset = (props: Props) => {
   useEffect(() => {
     loader.load(url, (texture) => {
       assets[url].loaded = true;
-      assets[url].data = texture;
+      assets[url].data = ref(texture);
     });
   }, [loader]);
