+++ src/core/loader/loaders/ModelAsset.tsx
@@ -1,22 +1,23 @@
-import { useEffect } from "react";
+import { useContext, useEffect } from "react";
 import { useGLTF } from "@react-three/drei";
 import { DRACO_URL } from "../../../services/constants";
-import { Assets } from "../../types/loading";
+import { LoadingContext } from "../../contexts";
+import { ref } from "valtio";
 
 type Props = {
   url: string;
-  assets: Assets;
 };
 
 const ModelAsset = (props: Props) => {
-  const { url, assets } = props;
+  const { url } = props;
+  const { assets } = useContext(LoadingContext);
 
   const gltf = useGLTF(url, DRACO_URL);
 
   useEffect(() => {
     if (gltf) {
       assets[url].loaded = true;
-      assets[url].data = gltf;
+      assets[url].data = ref(gltf);
     }
   }, [gltf]);
